refactor: image -> s3 패키지 변경
