+++ src/main/java/com/map/gaja/client/infrastructure/s3/S3FileService.java
@@ -1,4 +1,4 @@
-package com.map.gaja.client.infrastructure.image;
+package com.map.gaja.client.infrastructure.s3;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectMetadata;
@@ -12,7 +12,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.annotation.PostConstruct;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.UUID;
