+++ src/main/java/com/map/gaja/client/infrastructure/s3/S3UrlGenerator.java
@@ -1,4 +1,4 @@
-package com.map.gaja.client.infrastructure.image;
+package com.map.gaja.client.infrastructure.s3;
 
 import org.springframework.beans.factory.annotation.Value;
